feat: class PopupWithForm is ready, only one class left to complete
